refactor: Extract sorting logic into a separate function for improved readability and reusability
